requirements.txt
asgiref==3.3.4
attrs==19.3.0
audio-metadata==0.11.1
Babel==2.9.1
bidict==0.21.2
bitstruct==8.11.1
boolean.py==3.8
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
gitdb==4.0.7
GitPython==3.1.18
idna==2.10
more-itertools==8.8.0
pendulum==2.1.2
pkg-resources==0.0.0
pprintpp==0.4.0
psutil==5.8.0
PySimpleGUI==4.44.0
python-dateutil==2.8.1
python-vlc==3.0.12118
pytz==2021.1
pytzdata==2020.1
requests==2.25.1
shiboken2==5.15.2
six==1.16.0
smmap==4.0.0
sqlparse==0.4.1
tbm-utils==2.6.0
tkfilebrowser==2.3.2
typing-extensions==3.10.0.0
urllib3==1.26.5
watchdog==2.1.3
wrapt==1.12.1
